feat: fair randomness calculation
This implements a new algorithm for getting fortunes randomly, based on the
weighted sum of all fortunes, where the weight is the file size in bytes.

This prevents the problem of all files being uniformly distributed, leading to
small files being overrepresented, and thus often repeating fortunes from these
files, in a very unsatisfactory way.

Refs: #15
Signed-off-by: Christina Sørensen <christina@cafkafk.com>